Move main.go and add new features
The main.go file was moved to the DTAPI
