+++ src/DTAPI/TODO.md
@@ -1,0 +1,6 @@
+# Stuff to do to get feature parity
+- [ ] Accepts the /Proccess update
+- [ ] Does the /Update for DTService updating
+- [ ] Database connection + Handling
+
+# More stuff
